Refactor imports from `math`
