+++ src/tensortrax/math/_math_tensor.py
@@ -1,13 +1,6 @@
-r"""
- _
-| |                          ████████╗██████╗  █████╗ ██╗  ██╗
-| |_ ___ _ __  ___  ___  _ __╚══██╔══╝██╔══██╗██╔══██╗╚██╗██╔╝
-| __/ _ \ '_ \/ __|/ _ \| '__|  ██║   ██████╔╝███████║ ╚███╔╝
-| ||  __/ | | \__ \ (_) | |     ██║   ██╔══██╗██╔══██║ ██╔██╗
- \__\___|_| |_|___/\___/|_|     ██║   ██║  ██║██║  ██║██╔╝ ██╗
-                                ╚═╝   ╚═╝  ╚═╝╚═╝  ╚═╝╚═╝  ╚═╝
 """
-
+tensorTRAX: Math on (Hyper-Dual) Tensors with Trailing Axes.
+"""
 import numpy as np
 
 from .._tensor import Tensor, Δ, Δδ, broadcast_to, einsum, f, matmul, δ
