--- src/tensortrax/math/_special/__init__.py
@@ -1,31 +1,0 @@
-r"""
- _
-| |                          ████████╗██████╗  █████╗ ██╗  ██╗
-| |_ ___ _ __  ___  ___  _ __╚══██╔══╝██╔══██╗██╔══██╗╚██╗██╔╝
-| __/ _ \ '_ \/ __|/ _ \| '__|  ██║   ██████╔╝███████║ ╚███╔╝
-| ||  __/ | | \__ \ (_) | |     ██║   ██╔══██╗██╔══██║ ██╔██╗
- \__\___|_| |_|___/\___/|_|     ██║   ██║  ██║██║  ██║██╔╝ ██╗
-                                ╚═╝   ╚═╝  ╚═╝╚═╝  ╚═╝╚═╝  ╚═╝
-"""
-
-from ._special_tensor import (
-    ddot,
-    dev,
-    from_triu_1d,
-    from_triu_2d,
-    sym,
-    tresca,
-    triu_1d,
-    von_mises,
-)
-
-__all__ = [
-    "ddot",
-    "dev",
-    "from_triu_1d",
-    "from_triu_2d",
-    "sym",
-    "tresca",
-    "triu_1d",
-    "von_mises",
-]
